shopping screen refactor
