+++ shop/presentation/src/main/java/hu/tb/shopping/presentation/components/ShoppingEmptyScreen.kt
@@ -1,0 +1,42 @@
+package hu.tb.shopping.presentation.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.ShoppingCart
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ShoppingEmptyScreen() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Icon(
+            imageVector = Icons.Outlined.ShoppingCart,
+            tint = MaterialTheme.colorScheme.primary.copy(
+                alpha = 0.6f
+            ),
+            contentDescription = "shop cart empty icon"
+        )
+        Spacer(Modifier.height(16.dp))
+        Text(
+            text = "Your list is empty.",
+            style = MaterialTheme.typography.headlineMedium,
+            color = MaterialTheme.colorScheme.primary.copy(
+                alpha = 0.6f
+            )
+        )
+    }
+}
